Create carousel component
