+++ src/app/components/features/home/dedicated-products/dedicated-products.html
@@ -1,28 +1,15 @@
 <section class="flex flex-col items-center gap-8">
   <h1>{{'dedicated_products.header' | translate}}</h1>
-  <div class="relative flex justify-center w-full">
-    <button class="absolute right-8 top-1/2 bg-[#D3E0FF]/70 hover:bg-[#D3E0FF] active:bg-[#D3E0FF]/70 rounded-full p-4 transition-colors z-20" (click)="next()">
-      <img ngSrc="assets/arrow_forward_icon.svg" alt="arrow forward" width="24" height="24">
-    </button>
-    <button class="absolute left-8 top-1/2 bg-[#D3E0FF]/70 hover:bg-[#D3E0FF] active:bg-[#D3E0FF]/70 rounded-full p-4 transition-colors z-20" (click)="prev()">
-      <img ngSrc="assets/arrow_back_icon.svg" alt="arrow back" width="24" height="24">
-    </button>
-    <div class="relative w-11/12 overflow-hidden">
-      <ul #carouselTrack class="flex flex-row items-center w-full transition"
-          [style.transform]="'translateX(-' + translateX + 'px)'">
-        @for (product of products; track product.id){
-          <li class="w-1/1 sm:w-1/2 md:w-1/3 lg:w-1/4 xl:w-[20%] h-96 px-4 flex-shrink-0">
-            <div class="border-1 rounded-2xl w-full h-full">
-              <img [ngSrc]="product.image" [alt]="product.name" class="w-full h-[60%]" width="1" height="1">
-              <div class="flex flex-col gap-2 p-2">
-                <p class="text-3xl">{{ product.price }} $</p>
-                <p class="text-lg">{{ product.isNegotiable ? "negotiable" : "" }}</p>
-                <p class="text-lg">{{ product.name }}</p>
-              </div>
-            </div>
-          </li>
-        }
-      </ul>
-    </div>
-  </div>
+  <app-carousel [items]="products">
+    <ng-template let-product>
+      <div class="border-1 rounded-2xl w-full h-full">
+        <img [ngSrc]="product.image" [alt]="product.name" class="w-full h-[60%]" width="1" height="1">
+        <div class="flex flex-col gap-2 p-2">
+          <p class="text-3xl">{{ product.price }} $</p>
+          <p class="text-lg">{{ product.isNegotiable ? "negotiable" : "" }}</p>
+          <p class="text-lg">{{ product.name }}</p>
+        </div>
+      </div>
+    </ng-template>
+  </app-carousel>
 </section>
