+++ src/app/components/features/home/home.html
@@ -3,11 +3,6 @@
   <app-categories/>
   <div class="h-0.5 bg-[#AEAEAE] my-6"></div>
   <app-dedicated-products/>
-
-  <app-carousel>
-
-  </app-carousel>
-
   <app-baner/>
   <div class="flex flex-col md:flex-row md:items-start gap-4">
     <div class="md:w-2/3 w-full">
