+++ src/app/components/features/home/home.ts
@@ -4,9 +4,6 @@ import {Categories} from './categories/categories';
 import {DedicatedProducts} from './dedicated-products/dedicated-products';
 import {NewestProducts} from './newest-products/newest-products';
 import {Auctions} from './auctions/auctions';
-import {Carousel} from '../../shared/carousel/carousel';
-import {NgOptimizedImage} from '@angular/common';
-import {Product} from '../../../core/models/Product';
 
 @Component({
   selector: 'app-home',
@@ -16,8 +13,6 @@ import {Product} from '../../../core/models/Product';
     DedicatedProducts,
     NewestProducts,
     Auctions,
-    Carousel,
-    NgOptimizedImage
   ],
   templateUrl: './home.html',
   styleUrl: './home.css'
